+++ backend/src/oauth/provider/discord.provider.ts
@@ -1,5 +1,4 @@
 import { Injectable } from "@nestjs/common";
-import fetch from "node-fetch";
 import { ConfigService } from "../../config/config.service";
 import { OAuthCallbackDto } from "../dto/oauthCallback.dto";
 import { OAuthSignInDto } from "../dto/oauthSignIn.dto";
@@ -52,7 +51,7 @@ export class DiscordProvider implements OAuthProvider<DiscordToken> {
           this.config.get("general.appUrl") + "/api/oauth/callback/discord",
       }),
     });
-    const token: DiscordToken = await res.json();
+    const token = (await res.json()) as DiscordToken;
     return {
       accessToken: token.access_token,
       refreshToken: token.refresh_token,
@@ -102,10 +101,10 @@ export class DiscordProvider implements OAuthProvider<DiscordToken> {
       });
       const guilds = (await res.json()) as DiscordPartialGuild[];
       if (!guilds.some((guild) => guild.id === guildId)) {
-        throw new ErrorPageException("discord_guild_permission_denied");
+        throw new ErrorPageException("user_not_allowed");
       }
     } catch {
-      throw new ErrorPageException("discord_guild_permission_denied");
+      throw new ErrorPageException("user_not_allowed");
     }
   }
 }